Re-write day 15
The target prioritization code for day 15 was more complex than it
needed to be, tracking the target entity rather than the target
movement square.  Re-worked it by making all unreachable values now
usize::max_value(), and only considering adjacent squares to targets.
This means we need to be able to realize we should stay put with this
prioritization.